30th commit : rethinking of Freehand
I have completly rethought the Freehand functionning. I now divide the drawing space in a virtual grid of 10pixels sized squared acting as atoms saying wether they are occupied or not. This has improved a lot the efficiency of the path calculation algorithm, dividing by 100 the number of nodes to explore.

I also have rethought the displaying of a path : there is now a single "Rectangle" node following the path.